refactor: relocate music caption implementation
